Create a class named Horse that contains private data fields for the name, color, and birth year.
Include get and set methods for these fields. Additionally create in horse class a massage void

MethodHorse() that announce itself (“I am methodHorse in the Horse class”). Next, create a subclass

named RaceHorse, which contains an additional field that holds the number of races in which the

horse has competed and additional methods to get and set the new field. Write an application that

demonstrates the usage of objects of each class. Old classes should have a constructor that announces

itself for example in the class Horse the default constructor should announce: “I am the default

constructor of the class Horse”.

a. Save the files as Horse.java, RaceHorse.java, and DemoHorses.java.

b. In addition, with their path create a class diagram indicating all the classes with the

constructors with the instance methods.

c. Override the MethodHorse() of the superclass to announce itself into the following in the

derived class RaceHorse. The MethodHorse() now in the derived class will announce itself

(“I am methodHorse in the RaceHorse class”)