Create two classes, A and B, with default constructors (empty argument
lists) that announce themselves. Also create constructors with arguments

that they announce themselves and a string inserted

(i.e. public B(String s) {

 System.out.println("B(): " + s);)

Write a method void a() for A class, b() for B et.c that that announce

themselves where they belong(which class i.e. a() will println(“I am in A

class”)).Inherit a new class called C from A, and create a member of class B inside

C. Do not create a constructor for C. Create an object of class C and observe

the results.

Try to override the methods of A and B in C class. Observe your output and

write your comments.