Garbage Collection/public void finalize()
Create a class called Tank that can be filled and emptied, and has a death

condition that it must be empty when the object is cleaned up. Write a

finalize() that verifies this death condition.

In main(), test the possible scenarios that can occur when your Tank is used.

Your program must host the following:

static int counter;

 int id = counter++;

 boolean full = false;

The constructor of Tank() which prints itself and the number of Tank created i.e.Tank