Create a class named Vehicle that acts as a superclass for vehicle types. The
Vehicle class contains private variables for the number of wheels and the average

number of miles per gallon. The Vehicle class also contains a constructor with

integer arguments for the number of wheels and average miles per gallon, and a

toString() method that returns a String containing these values. Create two

subclasses, Car and MotorCycle, that extend the Vehicle class. Each subclass

contains a constructor that accepts the miles-per-gallon value as an argument

and forces the number of wheels to the appropriate value—2 for a MotorCycle

and 4 for a Car. Write a UseVehicle class to instantiate the two Vehicle objects and

display the objects’ values. Save the files as Vehicle.java, Car.java,

MotorCycle.java, and UseVehicle.java.