Create an inheritance hierarchy abstraction called Building with 2 abstract methods and
one concrete. From this, inherit a class called UniversityBuilding which is considered to

have at least 2 methods and the 2 abstract void methods. Put appropriate methods (of your

choice) in the base UniversityBuilding class. Then, create the classes EuropaBuilding and

MainBuilding which inherit and obey to all characteristics of Building and

UniversityBuilding (see the Diagram 1). In a TestClassUniversityBuilding, in the static

main( ) function, create EuropaBuilding and MainBuilding and upcast it to

UniversityBuilding, and demonstrate that all the methods declared work perfectly (of all

sides). In addition in the static main( ) function you should create an array of

UniversityBuilding and upcast it to EuropaBuilding and MainBuilding. The array should

be at least of 10 memory locations.