Create your own exception/ ScoreException & TestScore Write an application that displays a series of at least five student ID numbers
(that you have stored in an array) and asks the user to enter a numeric test

score for the student.

Create a ScoreException class, and throw a ScoreException for the

class if the user does not enter a valid score (less than or equal to 100). Catch

the ScoreException and then display an appropriate message. In

addition, store a 0 for the student's score. At the end of the application,

display all the student IDs and scores. Save the files as

ScoreExceptionJava and TestScoreJava.