Exercise 1: Blanket Class a. Create a class named Blanket with private fields for a blanket's size, color, material,
and price as well as for additional fields of your choice (all total fields 8). Include a

constructor that sets default values for the fields as Twin, white, cotton, and

0.00.

Include another 4 (four) non-default constructor of your choice initializing the

blanket's size, color, material, and price as well as other instance variables of your

choice (of any type).

In the class named Blanket include a set method for each one of the fields including

also the fields of your choice (all total fields 8). The method that sets size. adds 0

to the base price for a double blanket, 5 for a queen blanket, and 0 for a king.

The method that sets the material adds 0 to the price for wool and 5 to the

price for cashmere.

In other words, the price for a king-sized cashmere blanket is 15. Whenever the

size or material is invalid, reset the blanket to the default values. Include a toString()

method that returns a description of the blanket.

In addition, assign 2 final methods that in case of override in part (b) this may not

be allowed; additionally 2 protected methods of your choice along with their

accessors and their mutators in case they are accessed from different package.

Save the file as Blanket.java.

b. Create a derived class named ElectricBlanket from the class Blanket that includes

two additional fields: one for the number of heat settings and one for whether the

electric blanket has an automatic shutoff feature. Default values are one heat

3

setting and no automatic shutoff. Include get and set methods for the fields. Do not

allow the number of settings to be fewer than one or more than five; if it is, use the

default setting of 1. Add a .75 premium to the price if the blanket has the

automatic shutoff feature.

Also include a toString () method that calls the parent class toString() method and

combines the returned value with data about the new fields to return a complete

description of features.

Override two of the methods that are set in class blanket show that class

ElectricBlanket calls them in both superclasses versions and ElectricBlanket version

(derived class ).

Save the file as ElectricBlanket.java.

c. Create a complete and error free Java code application that declares a Blanket of

each type and demonstrates how the methods work. Your application should have

a user interface (UI menu) so that the user can select any part of the menu of your

application and create a specified Blanket. The menu selection items can be of your

choice.

Save the file as DemoBlankets.java.

Your program should clearly show characteristics Java characteristics like

inheritance polymorphism as well as upcast and downcast. Also, your classes

should use aggregation characteristics.

For all the above create a completely thorough class diagram indicating the

relations and associations between classes.